chore: separate integration and isolated tests
